// SalaryBooks-Front-development/src/app/models/TableFiilterOptions.ts
export default class TableFilterOptions {
  length: number = 10;
  searchkey: string = '';
  emp_status?: string = '';
  month?: string = '';
  year?: string = '';
  attendance_type?: string = '';
  emp_first_name?: string = '';
  emp_last_name?: string = '';
  emp_email_id?: string = '';
  department_id?: string = '';
  designation_id?: string = '';
  branch_id?: string = '';
  hod_id?: string = '';
  client_id?: string = '';
  advance_filter?: string = '';
  gender?: string = '';
  religion?: string = '';
  age_from?: string = '';
  age_to?: string = '';
  doj_from?: string = '';
  doj_to?: string = '';
  doe_from?: string = '';
  doe_to?: string = '';
  disbursement_frequency?: string = '';
  disbursement_type?: string = '';
}